backend/agent-a-summarizer/requirements.txt
# Agent A - Email Summarizer Requirements

# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Anthropic Claude
# anthropic>=0.8.0  # Replaced with free Google Gemini
google-generativeai>=0.3.0  # FREE Google Gemini API
langchain-google-genai>=0.0.5  # LangChain integration

# LangGraph for agent workflows
langgraph>=0.0.40
langchain>=0.1.0
langchain-anthropic>=0.1.0

# HTTP client for agent-to-agent communication
httpx>=0.26.0

# Celery for async tasks
celery>=5.3.4
redis>=5.0.0

# Google APIs
google-api-python-client>=2.116.0
google-auth>=2.27.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Shared module (local)
-e ../shared

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.1.0
